remove change of current working directory
Changing the current working directory conflicts with puppetfile mode.
As the working directory is not set back after all modules are
downloaded, the syncForgeToModuleDir() function will move them to
cache_tmp_dir/$module_dir, and not to ./$module_dir as expected.

Either the current working directory must be set back after the tar
file is extracted, which proves to be rather complicated as all
modules are processed asynchronously, and therefore could only be
done in a place not related to the actual change of the working
directory.

As it is better to not change the current working directory at all
the operations for the extracted files and directories should
include the absolute path for the target file, as implemented by
this patch.