add fallback attribute for Git modules
This tries to use different branches in a specific order until it finds a working branch.
Example:

```
mod 'puppetlabs-stdlib',
    :git => 'https://github.com/puppetlabs/puppetlabs-stdlib.git',
    :fallback => '4.889.x|foobar|master'
```

In this example g10k tries to use the branch `4.889.x`, which does not exist, then branch `foobar`, which also does not exist and finally branch `master`, which it can use.
All without failing or error messages.