Improve module deprecation check
The deprecation check queries the api each time also if the module is
already in the cache. This slows down the sync unnecessarily.

Improve this by reparsing the file when it is in the cache to continue
printing known deprecations and use the module TTL to refresh the
deprecation cache.

Closes gh-152